improvements on filters bar to allow user to have multiple filters selected
