+++ src/components/FiltersBar/index.tsx
@@ -2,30 +2,67 @@ import Colors from '@/constants/Colors';
 import { upperCaseString } from '@/src/util/productsUtil';
 import EvilIcons from '@expo/vector-icons/EvilIcons';
 import React from 'react';
-import { Text, TouchableOpacity, View } from 'react-native';
+import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import CategoryPicker from './CategoryPicker';
-import FilterButton from './FilterButton';
-import styles from './styles';
 
 interface FiltersBarProps {
   categories: string[];
-  selectedCategory: string | null;
-  setSelectedCategory: (category: string | null) => void;
+  selectedCategory: string[];
+  setSelectedCategory: (category: string) => void;
 };
 
 export default function FiltersBar({categories, selectedCategory, setSelectedCategory}:FiltersBarProps) {
     return (
         <View style={styles.filtersBarContainer}>
-            <FilterButton selectedCategory={selectedCategory} setSelectedCategory={setSelectedCategory} />
-
-            <CategoryPicker selectedCategory={selectedCategory} setSelectedCategory={setSelectedCategory} categories={categories} />
-
-            {selectedCategory && 
-                <TouchableOpacity style={styles.selectedFilter} onPress={() => {setSelectedCategory(null)}}>
-                    <Text>{upperCaseString(selectedCategory)}</Text>
-                    <EvilIcons name="close" size={16} color={Colors.blackText} />
-                </TouchableOpacity>
-            }
+            <View style={styles.buttonContainer}>
+                <CategoryPicker selectedCategory={selectedCategory} setSelectedCategory={setSelectedCategory} categories={categories} />
+            </View>
+            
+            <View style={styles.filtersContainer}>
+                {selectedCategory.map((category) => (
+                    <TouchableOpacity key={category} style={styles.selectedFilter} onPress={() => {setSelectedCategory(category)}}>
+                        <Text numberOfLines={1} ellipsizeMode="tail" style={styles.selectedFilterText}>{upperCaseString(category)}</Text>
+                        <EvilIcons name="close" size={16} color={Colors.blackText} />
+                    </TouchableOpacity>
+                ))  
+                }
+            </View>
         </View>
     );
 }
+
+const styles = StyleSheet.create({
+    filtersBarContainer: {
+        flexDirection: 'row',
+        flexWrap: 'nowrap',
+        paddingHorizontal: 10,
+        paddingVertical: 5,
+        marginBottom: 10,
+        backgroundColor: Colors.lightGrey,
+        alignItems: "center",
+    },
+    selectedFilter: {
+        backgroundColor: Colors.filterBackground,
+        padding: 6,
+        alignItems: "center",
+        margin: 2,
+        borderRadius: 4,
+        flexDirection: "row",
+        gap: 2,
+        maxWidth: 110
+    },
+    buttonContainer: {
+        justifyContent: 'center',
+        alignItems: 'center',
+        marginRight: 10,
+    },
+    filtersContainer: {
+        flex: 1,
+        flexDirection: 'row',
+        flexWrap: 'wrap',
+        alignItems: 'flex-start'
+    },
+    selectedFilterText: {
+        maxWidth: 80
+    }
+})
